Broaden types accepted by stirlings1 and stirlings2 functions
The functions stirlings1(n, k) and stirlings1(n, k) are currently (mid 2019) written to specify Int arguments for n and k, but the documentation states that they should optionally take a BigInt as well.  This can be fixed by changing them not to specify an Int for n and k.